Move Blocklist model into h.models package
As part of an effort to move all models into the h.models package, this
moves the Blocklist model out of h.badge and into h.models.
