+++ h/models/__init__.py
@@ -24,10 +24,10 @@ from h.features import models as features_models
 from h.groups import models as groups_models
 from h.nipsa import models as nipsa_models
 from h.notification import models as notification_models
-from h.badge import models as badge_models
 
 from h.models.activation import Activation
 from h.models.auth_client import AuthClient
+from h.models.blocklist import Blocklist
 from h.models.token import Token
 from h.models.user import User
 
@@ -48,7 +48,6 @@ __all__ = (
 )
 
 
-Blocklist = badge_models.Blocklist
 Feature = features_models.Feature
 FeatureCohort = features_models.FeatureCohort
 Group = groups_models.Group
